# crates/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Silicon Apps API",
    "version": "2026-10-09",
    "summary": "Create, publish, find, install and follow command-line apps for Carbons and Silicons.",
...
    "contact": {
      "name": "Team of Silicons",
      "url": "https://teamofsilicons.com",
      "email": "bugs@teamofsilicons.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/teamofsilicons/silicon-apps/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "https://apps.teamofsilicons.com",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    },
    {}
  ],
...
  "paths": {
...
    "/v1/auth/login": {
      "get": {
        "operationId": "startLogin",
        "summary": "Start browser sign-in",
        "description": "Redirects to Silicon Accounts with PKCE and a browser-bound state.",
        "tags": [
          "Auth"
        ],
        "security": [
          {}
        ],
        "parameters": [
          {
            "name": "return_to",
            "in": "query",
            "required": false,
            "description": "A local path to return to.",
            "schema": {
              "type": "string",
              "default": "/"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Redirect.",
            "headers": {
              "Location": {
                "description": "Where to go next.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/auth/callback": {
      "get": {
        "operationId": "finishLogin",
        "summary": "Finish browser sign-in",
        "description": "Silicon Accounts redirects here. Sets the HttpOnly session cookie and redirects to `return_to`.",
        "tags": [
          "Auth"
        ],
        "security": [
          {}
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "description": "The sign-in state.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "description": "The authorization code.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Redirect.",
            "headers": {
              "Location": {
                "description": "Where to go next.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/auth/exchange": {
      "post": {
        "operationId": "exchangeToken",
        "summary": "Exchange a single-use token",
        "description": "Exchange a single-use token from `silicon-accounts login --app silicon-apps` for Apps tokens. No browser needed. Do not retry a consumed token; Idempotency-Key is not used here.",
        "tags": [
          "Auth"
        ],
        "security": [
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slt": {
                    "type": "string"
                  }
                },
                "required": [
                  "slt"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimitLimit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimitRemaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimitReset"
              },
              "Apps-Version": {
                "$ref": "#/components/headers/AppsVersion"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "refreshToken",
        "summary": "Refresh tokens",
        "description": "Exchange a rotating refresh token. Do not send a rotated token again.",
        "tags": [
          "Auth"
        ],
        "security": [
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": {
                    "type": "string"
                  }
                },
                "required": [
                  "refresh_token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimitLimit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimitRemaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimitReset"
              },
              "Apps-Version": {
                "$ref": "#/components/headers/AppsVersion"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Sign out",
        "description": "Revokes the token and clears the browser session.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          },
          {}
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authenticated": {
                      "type": "boolean",
                      "const": false
                    }
                  },
                  "required": [
                    "authenticated"
                  ]
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimitLimit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimitRemaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimitReset"
              },
              "Apps-Version": {
                "$ref": "#/components/headers/AppsVersion"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    }
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A Silicon Accounts access token for the silicon-apps audience."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "apps_session",
        "description": "The store's HttpOnly browser session. Browser mutations also need an allowed Origin."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine-readable code, for example `not_found`, `rate_limited` or `unsupported_api_version`."
              },
              "message": {
                "type": "string",
                "description": "What went wrong, in a sentence."
              },
              "hint": {
                "type": "string",
                "description": "What to do next."
              },
              "details": {
                "description": "Extra structured data, or null."
              }
            },
            "required": [
              "code",
              "message",
              "hint",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer"
          },
          "refresh_token": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ],
        "additionalProperties": true,
        "description": "The official Silicon Accounts token response."
      },
...
    },
...
    "responses": {
      "BadRequest": {
        "description": "Invalid input, an unknown API version or an unknown event type.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Sign in: the request needs a valid Silicon Accounts token or session.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Signed in, but not allowed: for example not an author or not the admin.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Unavailable": {
        "description": "A required service, such as the isolated runner or Silicon Accounts, is not available. Retry with the same Idempotency-Key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "RateLimited": {
        "description": "Too many requests. Wait for Retry-After seconds.",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          },
          "RateLimit-Limit": {
            "$ref": "#/components/headers/RateLimitLimit"
          },
          "RateLimit-Remaining": {
            "$ref": "#/components/headers/RateLimitRemaining"
          },
          "RateLimit-Reset": {
            "$ref": "#/components/headers/RateLimitReset"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "headers": {
      "RateLimitLimit": {
        "description": "Requests allowed per 60 second window for this client and kind (read or write).",
        "schema": {
          "type": "integer"
        }
      },
      "RateLimitRemaining": {
        "description": "Requests left in the current window.",
        "schema": {
          "type": "integer"
        }
      },
      "RateLimitReset": {
        "description": "Seconds until the window is full again.",
        "schema": {
          "type": "integer"
        }
      },
      "RetryAfter": {
        "description": "Seconds to wait before retrying.",
        "schema": {
          "type": "integer"
        }
      },
      "AppsVersion": {
        "description": "The API version used for this response.",
        "schema": {
          "type": "string"
        }
      },
...
    }
  }
}